added more tnt tools

+++ src/main/java/com/jasonjat/testingmod/modpackets/ModPacketsC2S.java
@@ -4,6 +4,8 @@ import net.fabricmc.fabric.api.networking.v1.PacketSender;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.network.ServerPlayNetworkHandler;
@@ -13,7 +15,6 @@ public class ModPacketsC2S {
 
     public static void register() {
         ServerPlayNetworking.registerGlobalReceiver(ModPackets.GUI_PACKET, ModPacketsC2S::guiThing);
-        //ServerPlayNetworking.registerGlobalReceiver(ModPackets.EXPLOSIVE_ARROW_PACKET, ModPacketsC2S::spawnArrow);
     }
 
     private static void spawnArrow(MinecraftServer minecraftServer, ServerPlayerEntity serverPlayerEntity, ServerPlayNetworkHandler serverPlayNetworkHandler, PacketByteBuf packetByteBuf, PacketSender packetSender) {
@@ -23,10 +24,12 @@ public class ModPacketsC2S {
     }
 
     private static void guiThing(MinecraftServer minecraftServer, ServerPlayerEntity serverPlayerEntity, ServerPlayNetworkHandler serverPlayNetworkHandler, PacketByteBuf packetByteBuf, PacketSender packetSender) {
-        String red = packetByteBuf.readString();
-        System.out.println("This has reached the server! " + red);
-
         PlayerEntity player = (PlayerEntity) serverPlayerEntity;
-        player.kill();
+        String red = packetByteBuf.readString();
+        if (red.equals("Drop")) {
+            player.giveItemStack(new ItemStack(Items.DIAMOND, 64));
+        } else {
+            player.kill();
+        }
     }
 }
